requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Computer Vision
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
Pillow==10.2.0

# Face Recognition
insightface==0.7.3
onnxruntime==1.19.2
mediapipe==0.10.9
scikit-learn==1.3.2

# Deep Learning
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==42.0.0

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1

# HTTP Client (for camera API if needed)
httpx==0.26.0
requests==2.31.0

# Logging & Monitoring
python-json-logger==2.0.7

# Image Processing & Augmentation
albumentations==1.3.1
scikit-image==0.22.0

# Data Handling
pandas==2.1.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # for TestClient

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Multi-Agent System Dependencies
ultralytics==8.0.196  # YOLOv8 for detection
facenet-pytorch==2.5.3  # FaceNet model
timm==0.9.12  # Vision models
transformers==4.36.2  # CLIP, DINOv2, etc.
accelerate==0.25.0  # Hugging Face optimization

# For Jetson AGX Orin specific optimizations (install separately if needed)
# tensorrt
# pycuda
